fix: align signed varint encoding with official varint32/64 (ZigZag) definition
- VarInt.ReadInt32/WriteInt32/ReadInt64/WriteInt64 now ZigZag-encode/decode,
  matching Mojang bedrock-protocol-docs primitives (varint32/varint64 = ZigZag;
  varuint32/varuint64 = plain base-128). Fixes every signed-varint field in the
  generated packets at once (block positions, actions, dimensions, enums...).
- IPacket.Encode/Decode: packet-id prefix is varuint32 -> use unsigned read/write
  (ids > 63 would misalign under zigzag).
- ResourcePackClientResponsePacket: wire format corrected from real 1.26.40
  client captures (type index 0-3 + name string, Downloading carries pack list).

+++ Codec/IPacket.cs
@@ -15,7 +15,8 @@ namespace Protocol.Packets
 			bytes = data;
 			using (var mem = new MemoryStreamReader(data))
 			{
-				VarInt.ReadInt32(mem);
+				// Packet id prefix is an UNSIGNED varint (not the zigzag "varint32" primitive).
+				VarInt.ReadUInt32(mem);
 				Read(mem);
 			}
 		}
@@ -27,7 +28,7 @@ namespace Protocol.Packets
 				using (var mem = new MemoryStream())
 				{
 					var writer = new MemoryStreamWriter(mem);
-					VarInt.WriteInt32(mem, PacketId);
+					VarInt.WriteUInt32(mem,(uint)PacketId);
 					Write(writer);
 					mem.Flush();
 					bytes = mem.ToArray();
